removing fasthgt
